--- a/lighthouse-core/config/default.js
+++ b/lighthouse-core/config/default.js
@@ -4,7 +4,7 @@ module.exports = {
     "passName": "defaultPass",
     "recordNetwork": true,
     "recordTrace": true,
-    "pauseBeforeTraceEndMs": 1000,
+    "pauseBeforeTraceEndMs": 5000,
     "useThrottling": true,
     "gatherers": [
       "url",