--- a/lighthouse-core/lib/traces/pwmetrics-events.js
+++ b/lighthouse-core/lib/traces/pwmetrics-events.js
@@ -87,7 +87,7 @@ class Metrics {
         },
       },
       {
-        name: 'Time to Interactive',
+        name: 'Time to Interactive (vAlpha)',
         id: 'tti',
         getTs: auditResults => {
           const ttiExt = auditResults['time-to-interactive'].extendedInfo;
@@ -95,16 +95,16 @@ class Metrics {
         },
       },
       {
-        name: 'Time to Interactive (fMP Only)',
-        id: 'tti-fmp-only',
+        name: 'Time to Interactive (vAlpha non-visual)',
+        id: 'tti-non-visual',
         getTs: auditResults => {
           const ttiExt = auditResults['time-to-interactive'].extendedInfo;
           return ttiExt.value.timestamps.timeToInteractiveB;
         },
       },
       {
-        name: 'Time to Interactive (fMP Only - 5s)',
-        id: 'tti-fmp-only-5s',
+        name: 'Time to Interactive (vAlpha non-visual, 5s)',
+        id: 'tti-non-visual-5s',
         getTs: auditResults => {
           const ttiExt = auditResults['time-to-interactive'].extendedInfo;
           return ttiExt.value.timestamps.timeToInteractiveC;