First round of perf metric investigation, to be judged using #1936 :) Follow-up work includes adding in network reverse search methods for comparison.

- Adds two additional TTI candidates that don't rely on visual completeness.
	- First 500ms window after fMP where 90th percentile EIL is <50ms
	- First 5000ms window after fMP where 90th percentile EIL is <50ms (if at the end of trace, uses a minimum window of as low as 1000ms)
- Falls back to fMP when speedline fails to find screenshots in trace.
- Adds timing data to trace-of-tab computed artifact.