@patrickhulce 2017-04-03
> want to update pwmetrics-events so we can get these new TTIs plotted in the trace?

done

![image](https://cloud.githubusercontent.com/assets/2301202/24630350/353aadb6-1871-11e7-8b0f-c4917746905f.png)